# openapi.yaml
openapi: 3.0.3
info:
  title: Hateblog Backend API
  description: |
    hateblog リニューアル版のバックエンド API 仕様書。
    はてなブックマークのエントリー情報を提供するREST APIです。
  version: 1.0.0
  contact:
    name: Hateblog Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://hateblog.jp/api/v1
    description: 本番環境
security:
  - ApiKeyAuth: []
...
paths:
...
  /tags/{tag}/entries:
    get:
      tags:
        - tags
      summary: タグ別エントリー一覧取得
      description: |
        指定されたタグに紐づくエントリーを新着順で取得します。
        ブックマーク件数での閾値フィルタリングが可能です。
        タグ閲覧はサーバー側で自動的に記録されます（tag_view_historyテーブル）。
      operationId: getEntriesByTag
      parameters:
        - name: tag
          in: path
          description: タグ名
          required: true
          schema:
            type: string
            maxLength: 100
            example: "技術"
        - name: min_users
          in: query
          description: 最低ブックマーク件数（5/10/50/100/500/1000など）
          required: false
          schema:
            type: integer
            minimum: 0
            default: 5
            example: 10
        - name: limit
          in: query
          description: 取得件数
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
            example: 25
        - name: offset
          in: query
          description: オフセット（ページネーション用）
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
      responses:
        '200':
          description: 成功
          headers:
            X-Cache:
              $ref: '#/components/headers/CacheStatus'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntryListResponse'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: タグが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tags/trending:
    get:
      tags:
        - tags
      summary: トレンドタグ一覧取得
      description: |
        直近の人気エントリーで使用されているタグの一覧を取得します。
        指定時間内のエントリーを対象に、タグの出現頻度とスコアを集計して返します。
      operationId: getTrendingTags
      parameters:
        - name: hours
          in: query
          description: 集計対象時間（デフォルト 24時間）
          required: false
          schema:
            type: integer
            enum: [6, 12, 24, 48]
            default: 24
            example: 24
        - name: min_users
          in: query
          description: 対象エントリーの最低ブックマーク件数
          required: false
          schema:
            type: integer
            minimum: 0
            default: 5
            example: 10
        - name: limit
          in: query
          description: 取得件数
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
      responses:
        '200':
          description: 成功
          headers:
            X-Cache:
              $ref: '#/components/headers/CacheStatus'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendingTagsResponse'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tags/clicked:
    get:
      tags:
        - tags
      summary: クリックされたエントリーのタグ一覧取得
      description: |
        実際にクリックされたエントリーで使用されているタグの一覧を取得します。
        click_metricsテーブルを参照し、ユーザーの実際の興味関心を反映したタグを返します。
      operationId: getClickedTags
      parameters:
        - name: days
          in: query
          description: 集計対象日数（デフォルト 7日間）
          required: false
          schema:
            type: integer
            enum: [1, 7, 30]
            default: 7
            example: 7
        - name: limit
          in: query
          description: 取得件数
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
      responses:
        '200':
          description: 成功
          headers:
            X-Cache:
              $ref: '#/components/headers/CacheStatus'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClickedTagsResponse'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API認証キー。データベースに登録されたキーと照合します。
        ヘッダーに `X-API-Key-ID: your-api-key-id` と `X-API-Key: your-api-key` の両方を含めてリクエストしてください。

        例:
        ```
        X-API-Key-ID: 9f8e7d6c-5b4a-3c2d-1e0f-123456789abc
        X-API-Key: hb_live_1234567890abcdef1234567890abcdef
        ```
  responses:
    UnauthorizedError:
      description: 認証エラー - APIキーまたはAPIキーIDが無効または未提供
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "UNAUTHORIZED"
            message: "Missing API key or key ID"
  headers:
    CacheStatus:
      description: キャッシュ判定（HIT/MISS）
      schema:
        type: string
        enum: [HIT, MISS]
  schemas:
    Entry:
      type: object
      description: はてなブックマークエントリー
      required:
        - id
        - title
        - url
        - posted_at
        - bookmark_count
        - tags
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: エントリーID
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          description: エントリーのタイトル
          example: "Go言語による高速なWeb API開発"
        url:
          type: string
          format: uri
          description: エントリーのURL
          example: "https://example.com/article/golang-web-api"
        posted_at:
          type: string
          format: date-time
          description: 記事の投稿日時（ISO 8601形式）
          example: "2025-01-05T10:30:00Z"
        bookmark_count:
          type: integer
          minimum: 0
          description: はてなブックマーク件数
          example: 150
        excerpt:
          type: string
          nullable: true
          description: 記事本文の抜粋
          example: "Go言語を使ったWeb API開発の実践的なテクニックを紹介します..."
        subject:
          type: string
          nullable: true
          description: RSSフィードのsubject（画面非表示、内部利用）
          example: "テクノロジー"
        tags:
          type: array
          description: エントリーに紐づくタグ一覧
          items:
            $ref: '#/components/schemas/EntryTag'
          example:
            - tag_id: "123e4567-e89b-12d3-a456-426614174000"
              tag_name: "Go"
              score: 95
            - tag_id: "223e4567-e89b-12d3-a456-426614174001"
              tag_name: "プログラミング"
              score: 85
        favicon_url:
          type: string
          format: uri
          description: Favicon URL（FaviconプロキシAPI）
          example: "/api/v1/favicons?domain=example.com"
        created_at:
          type: string
          format: date-time
          description: レコード作成日時（ISO 8601形式）
          example: "2025-01-05T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: レコード更新日時（ISO 8601形式）
          example: "2025-01-05T12:00:00Z"
    EntryTag:
      type: object
      description: エントリーに紐づくタグ（スコア付き）
      required:
        - tag_id
        - tag_name
        - score
      properties:
        tag_id:
          type: string
          format: uuid
          description: タグID
          example: "123e4567-e89b-12d3-a456-426614174000"
        tag_name:
          type: string
          maxLength: 100
          description: タグ名
          example: "Go"
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: タグのスコア（Yahoo! キーフレーズ抽出APIから取得した重要度、0〜100）
          example: 95
    EntryListResponse:
      type: object
      description: エントリー一覧レスポンス
      required:
        - entries
        - total
        - limit
        - offset
      properties:
        entries:
          type: array
          description: エントリー一覧
          items:
            $ref: '#/components/schemas/Entry'
        total:
          type: integer
          minimum: 0
          description: 総件数（フィルタ適用後）
          example: 500
        limit:
          type: integer
          description: 1ページあたりの件数
          example: 25
        offset:
          type: integer
          description: オフセット
          example: 0
    TrendingTag:
      type: object
      description: トレンドタグ（出現回数付き）
      required:
        - id
        - name
        - occurrence_count
        - entry_count
      properties:
        id:
          type: string
          format: uuid
          description: タグID
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          maxLength: 100
          description: タグ名
          example: "Go"
        occurrence_count:
          type: integer
          minimum: 0
          description: 集計期間内での出現回数（エントリー数）
          example: 42
        entry_count:
          type: integer
          minimum: 0
          description: このタグに紐づく総エントリー数
          example: 1234
    TrendingTagsResponse:
      type: object
      description: トレンドタグ一覧レスポンス
      required:
        - tags
        - hours
        - total
      properties:
        tags:
          type: array
          description: トレンドタグ一覧
          items:
            $ref: '#/components/schemas/TrendingTag'
        hours:
          type: integer
          description: 集計対象時間
          example: 24
        total:
          type: integer
          minimum: 0
          description: 該当タグ総数
          example: 50
    ClickedTag:
      type: object
      description: クリックされたタグ（クリック数付き）
      required:
        - id
        - name
        - click_count
        - entry_count
      properties:
        id:
          type: string
          format: uuid
          description: タグID
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          maxLength: 100
          description: タグ名
          example: "Go"
        click_count:
          type: integer
          minimum: 0
          description: 集計期間内でのクリック数
          example: 156
        entry_count:
          type: integer
          minimum: 0
          description: このタグに紐づく総エントリー数
          example: 1234
    ClickedTagsResponse:
      type: object
      description: クリックされたタグ一覧レスポンス
      required:
        - tags
        - days
        - total
      properties:
        tags:
          type: array
          description: クリックされたタグ一覧
          items:
            $ref: '#/components/schemas/ClickedTag'
        days:
          type: integer
          description: 集計対象日数
          example: 7
        total:
          type: integer
          minimum: 0
          description: 該当タグ総数
          example: 30
...
    ErrorResponse:
      type: object
      description: エラーレスポンス
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: エラーコード
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: エラーメッセージ
          example: "Invalid parameter: date format must be YYYYMMDD"
        details:
          type: object
          nullable: true
          description: 詳細情報（オプション）
          additionalProperties: true
          example:
            field: "date"
            value: "20251301"
            constraint: "month must be 1-12"